third_party/WebKit/LayoutTests/fast/dom/shadow/tabstop-property-expected.txt

This is basic behavior test for tabStop attribute.

Testing tabStop property and attribute
Test tabStop normal assignment behavior
PASS div.tabStop is false
PASS div.tabStop is false
PASS div.getAttribute('tabstop') is "false"
PASS div.tabStop is false
PASS div.hasAttribute('tabstop') is false
PASS div.tabStop is true
PASS div.getAttribute('tabstop') is "true"
PASS div.tabStop is false
PASS div.hasAttribute('tabstop') is false
Test tabStop value implicitly set by tabindex property
PASS div.tabStop is true
PASS div.tabStop is false
PASS div.tabStop is true
PASS div.tabStop is false
Test explicit tabStop not overridable by tabindex attribute
PASS div.tabStop is false
PASS div.getAttribute('tabstop') is "false"
PASS div.tabStop is true
PASS div.getAttribute('tabstop') is "true"
PASS div.tabStop is false
PASS div.getAttribute('tabstop') is "false"
PASS div.tabStop is false
PASS div.hasAttribute('tabstop') is false
Test tabStop change with explicit tabindex change
PASS div.tabStop is true
PASS div.tabStop is false
PASS div.getAttribute('tabstop') is "false"
PASS div.tabStop is false
PASS div.tabStop is true
PASS div.getAttribute('tabstop') is "true"
Test tabStop with invalid value
PASS div.tabStop is false
PASS div.getAttribute('tabstop') is "false"
PASS div.tabStop is true
PASS div.getAttribute('tabstop') is "invalid"
PASS div.tabStop is true
PASS div.getAttribute('tabstop') is "true"
PASS div.tabStop is false
PASS div.getAttribute('tabstop') is "invalid"
Test finished.
PASS successfullyParsed is true

TEST COMPLETE
